fix: move is-plain-object for es5 compilation
