+++ packages/next/shared/lib/is-plain-object.ts
